+++ photutils/psf_matching/__init__.py
@@ -1,0 +1,8 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+Subpackage containing tools for generating kernels for matching point
+spread functions.
+"""
+
+from .fourier import *  # noqa: F401, F403
+from .windows import *  # noqa: F401, F403